adding R script to get average distance plots
Added original working script to get the plots of average bead-cell distance or fold change. Works with fgf24 or tbx5 but does have on value that is hard coded that I would like to get rid of. you have to change 20 to 30 or vice versa in calculating distances based on number of cells tracked.